// src/components/RateReports/enums.ts
/* eslint-disable no-unused-vars */

export enum DataItemName {
  TANUMBER = 'TANUMBER',
  SHARECLASSID = 'SHARECLASSID',
  SHARECLASSNAME = 'SHARECLASSNAME',
  FUNDNAME = 'FUNDNAME',
  FUNDNAME_SHARECLASSNAME = 'FUNDNAME_SHARECLASSNAME',
  DAILYWEIGHTEDAVERAGEMATURITY = 'DAILYWEIGHTEDAVERAGEMATURITY',
  TOTALRETURNSASOFDATE = 'TOTALRETURNSASOFDATE',
  NETFUNDASSETS = 'NETFUNDASSETS',
  CUSTOMTRADINGDEADLINES = 'CUSTOMTRADINGDEADLINES',
  FACTSHEETURL = 'FACTSHEETURL',
  SUMMARYPROSPECTUSURL = 'SUMMARYPROSPECTUSURL',
  PROSPECTUSURL = 'PROSPECTUSURL',
  ANNUALREPORTURL = 'ANNUALREPORTURL',
  DAILYFACTOR = 'DAILYFACTOR',
  ONEDAYYIELD = 'ONEDAYYIELD',
  SEVENDAYCURRENTYIELD = 'SEVENDAYCURRENTYIELD',
  SEVENDAYPREWAIVERYIELD = 'SEVENDAYPREWAIVERYIELD',
  SEVENDAYEFFECTIVEYIELD = 'SEVENDAYEFFECTIVEYIELD',
  THIRTYDAYCURRENTYIELD = 'THIRTYDAYCURRENTYIELD',
  THIRTYDAYEFFECTIVEYIELD = 'THIRTYDAYEFFECTIVEYIELD',
  THIRTYDAYSECYIELD = 'THIRTYDAYSECYIELD',
  MININITIALINVESTMENT = 'MININITIALINVESTMENT',
  GROSSEXPENSERATIO = 'GROSSEXPENSERATIO',
  NETEXPENSERATIO = 'NETEXPENSERATIO',
  TICKER = 'TICKER',
  CUSIP = 'CUSIP',
  AVERAGEEFFECTIVEDURATION = 'AVERAGEEFFECTIVEDURATION',
  DAILYNAV = 'DAILYNAV',
  MONTHLYONEYEARCUMULATIVE = 'MONTHLYONEYEARCUMULATIVE',
  MONTHLYFIVEYEARANNUALIZED = 'MONTHLYFIVEYEARANNUALIZED',
  MONTHLYTENYEARANNUALIZED = 'MONTHLYTENYEARANNUALIZED',
  MONTHLYSINCEINCEPTION = 'MONTHLYSINCEINCEPTION',
}